requirements.txt
git+https://github.com/openai/whisper.git
transformers
ffmpeg-python==0.2.0
gradio
yt-dlp